recommendUtils.js for recommended functions in backend, userActionRoutes.js for liking, saving articles and updating lastRead and preferredTopics
