+++ backend/index.js
@@ -10,10 +10,13 @@ const Replicate = require("replicate");
 const Groq = require("groq-sdk");
 const natural = require('natural');
 const stopword = require('stopword');
+bodyParser = require('body-parser');
+
 const { scrapeArticle } = require('./scraper');
-bodyParser = require('body-parser')
+const { getAllPreferredTopics, generateFrequencyDictionary, calculateUserSimilarity, findSimilarUsers, recommendArticles } = require('./recommendUtils');
 const authRoutes = require('./authRoutes');
 const pulsecheckRoutes = require('./pulsecheckRoutes')
+const userActionRoutes = require('./userActionRoutes')
 
 
 const prisma = new PrismaClient();
@@ -28,8 +31,11 @@ const tokenizer = new natural.WordTokenizer();
 app.use(bodyParser.json());
 app.use(express.json());
 app.use(cors()); 
+
+
 app.use('/auth', authRoutes);
 app.use('/llama3', pulsecheckRoutes);
+app.use('/update', userActionRoutes);
 
 
 
@@ -37,52 +43,6 @@ app.listen(PORT, () => {
   console.log(`Server running on http://localhost:${PORT}`);
 });
 
-
-// update preferredTopics, lastRead
-app.patch('/api/users/:id', async (req, res) => {
-  const { id } = req.params;
-  const { preferredTopics, lastRead } = req.body;
-
-  if (!id) {
-      return res.status(400).json({ error: 'User ID is required' });
-  }
-
-  // Ensure the id is an integer
-  const userId = parseInt(id);
-  if (isNaN(userId)) {
-      return res.status(400).json({ error: 'Invalid user ID' });
-  }
-
-  try {
-      const data = {};
-
-      if (preferredTopics !== undefined) {
-          if (!Array.isArray(preferredTopics)) {
-              return res.status(400).json({ error: 'preferredTopics should be an array' });
-          }
-          data.preferredTopics = preferredTopics;
-      }
-
-      if (lastRead !== undefined) {
-          data.lastRead = lastRead;
-      }
-
-      const user = await prisma.user.update({
-          where: { id: userId },
-          data,
-      });
-
-      res.status(200).json({ message: 'User updated successfully!', user });
-  } catch (error) {
-      console.error('Error updating user:', error);
-
-      if (error.code === 'P2025') { // Prisma specific error code for record not found
-          return res.status(404).json({ error: 'User not found' });
-      }
-
-      res.status(500).json({ error: 'Failed to update user.' });
-  }
-});
 
 
 const fetchAndStoreArticles = async () => {
@@ -187,91 +147,6 @@ app.post('/api/detect-ai-content-hf', async (req, res) => {
 });
 
 
-// Like an article
-app.post('/api/articles/:articleId/like', async (req, res) => {
-    const { articleId } = req.params;
-    const { userId } = req.body;
-  
-    try {
-      // Update the likes count in the Article model
-      await prisma.article.update({
-        where: { id: parseInt(articleId) },
-        data: {
-          likes: {
-            increment: 1,
-          },
-        },
-      });
-  
-      // Update the liked articles in the User model
-      await prisma.user.update({
-        where: { id: parseInt(userId) },
-        data: {
-          liked: {
-            push: parseInt(articleId),
-          },
-        },
-      });
-  
-      // Create an interaction
-      await prisma.interaction.create({
-        data: {
-          userId: parseInt(userId),
-          articleId: parseInt(articleId),
-          liked: true,
-        },
-      });
-  
-      res.status(200).json({ message: 'Article liked successfully!' });
-    } catch (error) {
-      console.error('Error liking article:', error);
-      res.status(500).json({ error: 'Error liking article' });
-    }
-  });
-  
-
-// Save an article
-app.post('/api/articles/:articleId/save', async (req, res) => {
-  const { articleId } = req.params;
-  const { userId } = req.body;
-
-  try {
-    // Update the saves count in the Article model
-    await prisma.article.update({
-      where: { id: parseInt(articleId) },
-      data: {
-        saves: {
-          increment: 1,
-        },
-      },
-    });
-
-    // Update the saved articles in the User model
-    await prisma.user.update({
-      where: { id: parseInt(userId) },
-      data: {
-        saved: {
-          push: parseInt(articleId),
-        },
-      },
-    });
-
-    // Create an interaction
-    await prisma.interaction.create({
-      data: {
-        userId: parseInt(userId),
-        articleId: parseInt(articleId),
-        saved: true,
-      },
-    });
-
-    res.status(200).json({ message: 'Article saved successfully!' });
-  } catch (error) {
-    console.error('Error saving article:', error);
-    res.status(500).json({ error: 'Error saving article' });
-  }
-});
-
 
 // view liked or saved articles 
 app.get('/api/interactedArticles', async (req, res) => {
@@ -327,7 +202,6 @@ const getMostRecentArticlesByKeywords = async (keywords, limit = 2) => {
 };
 
 
-
 // get article headlines for marquee
 app.get(`/relevant-articles`, async (req, res) => {
   const { topics } = req.query;
@@ -348,41 +222,6 @@ app.get(`/relevant-articles`, async (req, res) => {
 
 
 
-
-
-
-// stuff for recommended algorithm: 
-
-// fetch preferred topics across all users
-const getAllPreferredTopics = async () => {
-  try {
-    const users = await prisma.user.findMany({
-      select: {
-        preferredTopics: true,
-      },
-    });
-
-    // transform and flatten an array of arrays into a single array.
-    return users.flatMap(user => user.preferredTopics); 
-  } catch (error) {
-    console.error('Error fetching preferred topics:', error);
-  }
-};
-
-
-// generate a frequency dictionary of the topics that users are interested in. 
-const generateFrequencyDictionary = (topics) => {
-  const frequencyDict = {};
-
-  topics.forEach(topic => {
-    if (!frequencyDict[topic]) {
-      frequencyDict[topic] = 0;
-    }
-    frequencyDict[topic] += 1;
-  });
-
-  return frequencyDict;
-};
 
 
 // fetch and cache articles based on the topics in this frequency dictionary
@@ -430,87 +269,34 @@ const fetchAndCacheArticlesByTopics = async (topics, limit = 3) => {
 
 
 const scheduleArticleFetching = () => {
-cron.schedule('*/30 * * * *', async () => {
-  await fetchAndStoreArticles();
+  // Schedule the article fetching and caching to run every 30 minutes
+  cron.schedule('*/30 * * * *', async () => {
+    // Fetch and store the latest articles
+    await fetchAndStoreArticles();
 
-  const preferredTopics = await getAllPreferredTopics();
-  const frequencyDict = generateFrequencyDictionary(preferredTopics);
-  const sortedTopics = Object.keys(frequencyDict).sort((a, b) => frequencyDict[b] - frequencyDict[a]);
-  const topNTopics = sortedTopics.slice(0, 5);
-  await fetchAndCacheArticlesByTopics(topNTopics);
-});
+    // Retrieve all user preferred topics
+    const preferredTopics = await getAllPreferredTopics();
+    // Generate a frequency dictionary for the preferred topics
+    const frequencyDict = generateFrequencyDictionary(preferredTopics);
+    const sortedTopics = Object.keys(frequencyDict).sort((a, b) => frequencyDict[b] - frequencyDict[a]);
+    const topNTopics = sortedTopics.slice(0, 5);
+    // Fetch and cache articles for the top 5 topics
+    await fetchAndCacheArticlesByTopics(topNTopics);
+  });
 
-fetchAndStoreArticles();
-(async () => {
-  const preferredTopics = await getAllPreferredTopics();
-  const frequencyDict = generateFrequencyDictionary(preferredTopics);
-  const sortedTopics = Object.keys(frequencyDict).sort((a, b) => frequencyDict[b] - frequencyDict[a]);
-  const topNTopics = sortedTopics.slice(0, 5);
-  await fetchAndCacheArticlesByTopics(topNTopics);
-})();
+  // Immediately fetch and store the latest articles
+  fetchAndStoreArticles();
+  (async () => {
+    const preferredTopics = await getAllPreferredTopics();
+    const frequencyDict = generateFrequencyDictionary(preferredTopics);
+    const sortedTopics = Object.keys(frequencyDict).sort((a, b) => frequencyDict[b] - frequencyDict[a]);
+    const topNTopics = sortedTopics.slice(0, 5);
+
+    await fetchAndCacheArticlesByTopics(topNTopics);
+  })();
 };
 
 scheduleArticleFetching();
-
-
-// calculate a similarity score between 2 users based on common likes, saves and preferredTopics
-const calculateUserSimilarity = (user1, user2) => {
-  const commonLikes = user1.liked.filter(articleId => user2.liked.includes(articleId));
-  const commonSaves = user1.saved.filter(articleId => user2.saved.includes(articleId));
-  const commonSaveToLike = user1.saved.filter(articleId => user2.liked.includes(articleId));
-  const commonLikeToSave = user1.liked.filter(articleId => user2.saved.includes(articleId));
-  const commonTopics = user1.preferredTopics.filter(topic => user2.preferredTopics.includes(topic));
-
-  const weight1 = 3; // Weight for common likes - middle weight because liked is more relevant than topics chosen when setting up profile
-  const weight2 = 4; // Weight for common saves - most weight because saved holds gravity
-  const weight3 = 2; // Weight for common topics - least wegiht because initial interests (can be changed but matters less than the articles)
-  const weight4 = 2; 
-  const weight5 = 2; 
-
-  const similarityScore = (weight1 * commonLikes.length) + (weight2 * commonSaves.length) + (weight3 * commonTopics.length) + (weight4 * commonLikeToSave.length) + (weight5 * commonSaveToLike.length);
-
-  return similarityScore;
-};
-
-
-// find similar users
-const findSimilarUsers = (targetUser, allUsers, similarityThreshold) => {
-  return allUsers.filter(user => {
-    if (user.id !== targetUser.id) {
-      const similarityScore = calculateUserSimilarity(targetUser, user);
-      return similarityScore > similarityThreshold;
-    }
-    return false;
-  });
-};
-
-
-// recommend articles based on what similar users liked, saved that target user has not interacted with yet. 
-const recommendArticles = (targetUser, similarUsers, allArticles) => {
-  const articleScores = {};
-
-  // for each similar user
-  similarUsers.forEach(similarUser => {
-    // loop through that user's interactions
-    similarUser.interactions.forEach(interaction => {
-      if ((interaction.liked || interaction.saved) &&
-          !targetUser.liked.includes(interaction.articleId) &&
-          !targetUser.saved.includes(interaction.articleId)) {
-        if (!articleScores[interaction.articleId]) {
-          articleScores[interaction.articleId] = 0;
-        }
-        // if article has been liked / saved by similar user and not by target users, add to its score
-        articleScores[interaction.articleId] += 1;
-      }
-    });
-  });
-
-  // sort articles from highest to lowest score
-  const sortedArticleIds = Object.keys(articleScores).sort((a, b) => articleScores[b] - articleScores[a]);
-
-  // get articleIds of the articles so they can be fetched to target user 
-  return sortedArticleIds.map(articleId => allArticles.find(article => article.id === parseInt(articleId)));
-};
 
 
 // create API endpoint to get recommended articles. 
@@ -589,6 +375,7 @@ async function updateArticleKeywords() {
       console.error('Error updating keywords:', error);
   } 
 }; 
+
 updateArticleKeywords();
