+++ frontend/src/components/TopicSelector.jsx
@@ -17,7 +17,7 @@ const TopicSelector = ({ user, setTopics }) => {
 
     const handleSubmit = async () => {
         try {
-            await axios.patch(`${import.meta.env.VITE_BACKEND_URL}/api/users/${user.id}`, { preferredTopics: selectedTopics });
+            await axios.patch(`${import.meta.env.VITE_BACKEND_URL}/update/users/${user.id}`, { preferredTopics: selectedTopics });
             setTopics(selectedTopics);
             navigate(`/profile/${user.id}`);
         } catch (error) {
